Rename news block to news-banner
Initially set empty news banner, then update is async
