+++ test/blocks/news-banner/news-banner.plain.html
@@ -1,0 +1,4 @@
+  <div>
+    <div class="news-banner block" data-block-name="news-banner" data-block-status="loading">
+    </div>
+  </div>
