--- test/blocks/news/news.plain.html
@@ -1,2 +1,0 @@
-<div class="news block" data-block-name="news" data-block-status="loading">
-</div>
